Refactor TUI app.rs into focused submodules (#201)
Split the 3,968-line app.rs into 8 focused files under app/ using
multiple impl App blocks. No behavioral changes — purely structural.
